requirements.txt
altgraph==0.17.4
blinker==1.9.0
bottle==0.13.4
cffi==2.0.0
click==8.2.1
clr_loader==0.2.7.post0
colorama==0.4.6
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pefile==2023.2.7
proxy_tools==0.1.0
pycparser==2.23
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.8
pythonnet==3.0.5
pywebview==6.0
pywin32-ctypes==0.2.3
typing_extensions==4.15.0
Werkzeug==3.1.3